+++ NuGet.config
@@ -15,8 +15,4 @@
   <activePackageSource>
     <add key="All" value="(Aggregate source)" />
   </activePackageSource>
-  <disabledPackageSources>
-    <add key="nugetbuild" value="true" />
-    <add key="dotnetcore" value="true" />
-  </disabledPackageSources>
 </configuration>
