+++ src/NuGetGallery/Services/PackageFileService.cs
@@ -19,7 +19,7 @@ namespace NuGetGallery
         {
             _fileStorageService = fileStorageService;
         }
-        
+
         public Task<ActionResult> CreateDownloadPackageActionResultAsync(Uri requestUrl, Package package)
         {
             var fileName = BuildFileName(package);
