+++ src/NuGetGallery/Web.config
@@ -24,9 +24,9 @@
     <!-- DEPLOYMENT SPECIFIC -->
     <!-- ******************* -->
     <!-- These should change on every deployment (to rotate credentials, etc.) -->
-    <add key="Gallery.StorageType" value="AzureStorage" />
+    <add key="Gallery.StorageType" value="FileSystem" />
     <!-- The storage type to use. Can be FileSystem (default) or AzureStorage -->
-    <add key="Gallery.AzureStorageConnectionString" value="UseDevelopmentStorage=true" />
+    <add key="Gallery.AzureStorageConnectionString" value="" />
     <!-- The connection string for the Azure Storage Account used for Package Storage IF Gallery.StorageType is AzureStorage -->
     <add key="Gallery.AzureStorageReadAccessGeoRedundant" value="false" />
     <!-- If the storage account has to fall-back to a secondary (when Read Access Geo Redundant is enabled in Azure storage), change Gallery.AzureStorageReadAccessGeoRedundant to true -->
